@@ -11,38 +11,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.0.6",
-    "@angular/common": "~11.0.6",
-    "@angular/compiler": "~11.0.6",
-    "@angular/core": "~11.0.6",
-    "@angular/forms": "~11.0.6",
-    "@angular/platform-browser": "~11.0.6",
-    "@angular/platform-browser-dynamic": "~11.0.6",
-    "@angular/router": "~11.0.6",
+    "@angular/animations": "~13.1.2",
+    "@angular/common": "~13.1.2",
+    "@angular/compiler": "~13.1.2",
+    "@angular/core": "~13.1.2",
+    "@angular/forms": "~13.1.2",
+    "@angular/platform-browser": "~13.1.2",
+    "@angular/platform-browser-dynamic": "~13.1.2",
+    "@angular/router": "~13.1.2",
     "bootstrap": "^3.3.7",
-    "rxjs": "~6.6.3",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.2"
+    "rxjs": "^7.5.2",
+    "tslib": "^2.3.1",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.6",
-    "@angular/cli": "~11.0.6",
-    "@angular/compiler-cli": "~11.0.6",
-    "@angular/language-service": "~11.0.6",
+    "@angular-devkit/build-angular": "~13.1.3",
+    "@angular/cli": "~13.1.3",
+    "@angular/compiler-cli": "~13.1.2",
+    "@angular/language-service": "~13.1.2",
+    "@types/jasmine": "^3.10.3",
+    "@types/jasminewd2": "^2.0.10",
     "@types/node": "^12.11.1",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.3",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.3.11",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
-    "ts-node": "~7.0.0",
+    "ts-node": "^10.4.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.0.5"
+    "typescript": "~4.5.4"
   }
 }
