@@ -4,14 +4,14 @@ import { AbstractControl, ValidationErrors, ValidatorFn } from '@angular/forms';
 import { Observable } from 'rxjs';
 
 export class CustomValidators {
-  static serviceLevel(c: AbstractControl): ValidationErrors | null {
+  static serviceLevel({ value }: AbstractControl): ValidationErrors | null {
     console.log('Validator: serviceLevel is called');
-    return checkServiceLevel(c);
+    return checkServiceLevel(value);
   }
 
   static serviceLevelRange(min: number, max: number): ValidatorFn {
-    return (c: AbstractControl): ValidationErrors | null => {
-      return checkServiceLevel(c, min, max);
+    return ({ value }: AbstractControl): ValidationErrors | null => {
+      return checkServiceLevel(value, min, max);
     };
   }
 
@@ -50,14 +50,10 @@ export class CustomValidators {
   }
 }
 
-export function checkServiceLevel(
-  c: AbstractControl,
-  min: number = 1,
-  max: number = 5
-): ValidationErrors | null {
+export function checkServiceLevel(value: number, min: number = 1, max: number = 5): ValidationErrors | null {
   if (
-    c.value !== undefined &&
-    (Number.isNaN(c.value) || c.value < min || c.value > max)
+    value !== undefined &&
+    (Number.isNaN(value) || value < min || value > max)
   ) {
     return {
       serviceLevel: true
